Keep video output in separate header, as in ESP_8_BIT
